+++ src/main/java/net/sf/jsqlparser/expression/ExpressionVisitor.java
@@ -44,7 +44,7 @@ import net.sf.jsqlparser.expression.operators.relational.Matches;
 import net.sf.jsqlparser.expression.operators.relational.MinorThan;
 import net.sf.jsqlparser.expression.operators.relational.MinorThanEquals;
 import net.sf.jsqlparser.expression.operators.relational.NotEqualsTo;
-import net.sf.jsqlparser.expression.operators.relational.PostgreSQLFromForExpression;
+import net.sf.jsqlparser.expression.operators.relational.SubstringExpression;
 import net.sf.jsqlparser.expression.operators.relational.RegExpMatchOperator;
 import net.sf.jsqlparser.expression.operators.relational.RegExpMySQLOperator;
 import net.sf.jsqlparser.schema.Column;
@@ -166,6 +166,6 @@ public interface ExpressionVisitor {
     
     void visit(OracleHint hint);
     
-    void visit(PostgreSQLFromForExpression postgreSQLFromForExpression);
+    void visit(SubstringExpression substringExpression);
     
 }
