+++ src/main/java/net/sf/jsqlparser/expression/operators/relational/SubstringExpression.java
@@ -1,0 +1,79 @@
+/*
+ * #%L
+ * JSQLParser library
+ * %%
+ * Copyright (C) 2004 - 2015 JSQLParser
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 2.1 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Lesser Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Lesser Public
+ * License along with this program.  If not, see
+ * <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ * #L%
+ */
+package net.sf.jsqlparser.expression.operators.relational;
+
+import net.sf.jsqlparser.expression.Expression;
+import net.sf.jsqlparser.expression.ExpressionVisitor;
+
+public class SubstringExpression implements Expression {
+
+	private Expression sourceExpression;
+	private Expression fromExpression;
+	private Expression forExpression;
+
+	public Expression getSourceExpression() {
+		return sourceExpression;
+	}
+
+	public void setSourceExpression(Expression sourceExpression) {
+		this.sourceExpression = sourceExpression;
+	}
+
+	public Expression getFromExpression() {
+		return fromExpression;
+	}
+
+	public void setFromExpression(Expression fromExpression) {
+		this.fromExpression = fromExpression;
+	}
+
+	public boolean hasFromExpression() {
+		return fromExpression != null;
+	}
+
+	public Expression getForExpression() {
+		return forExpression;
+	}
+
+	public void setForExpression(Expression forExpression) {
+		this.forExpression = forExpression;
+	}
+
+	public boolean hasForExpression() {
+		return forExpression != null;
+	}
+
+	@Override
+	public void accept(ExpressionVisitor expressionVisitor) {
+		expressionVisitor.visit(this);
+	}
+
+	@Override
+	public String toString() {
+		return
+				"SUBSTRING(" + sourceExpression
+				+ (fromExpression != null ? " FROM " + fromExpression : "")
+				+ (forExpression != null ? " FOR " + forExpression : "")
+				+ ")";
+	}
+
+}
